requirements.txt
# Core
anthropic>=0.40.0
PySide6>=6.6.0
PySide6-Addons>=6.6.0

# Document Processing
PyMuPDF>=1.24.0
python-docx>=1.1.0
openpyxl>=3.1.0
chardet>=5.0

# Token & Cost
tiktoken>=0.7.0

# Security
keyring>=25.0

# UI Rendering
markdown>=3.5
pymdown-extensions>=10.0
Pygments>=2.17

# Image
Pillow>=10.0

# Utilities
darkdetect>=0.8.0
httpx[socks]>=0.27.0
